Move MergeMaps to util
